% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/waffle.R
\name{as_rcdimple}
\alias{as_rcdimple}
\title{Turn a ggplot waffle chart object into an htmlwidget}
\usage{
as_rcdimple(wf, height = NULL, width = NULL)
}
\arguments{
\item{wf}{waffle chart ggplot2 object}

\item{height}{height of the resultant \code{htmlwidget}}

\item{width}{width of the resultant \code{htmlwidget}}
}
\description{
Takes the output from \code{waffle} and turns it into an \code{rcdimple}
\code{htmlwidget}.  For this to work, you will need to install
\code{rcdimple} with \code{devtools::install_github("timelyportfolio/rcdimple")}.
}

