% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-pictogram.R
\name{draw_key_pictogram}
\alias{draw_key_pictogram}
\title{Legend builder for pictograms}
\usage{
draw_key_pictogram(data, params, size)
}
\arguments{
\item{data, params, size}{legend key things}
}
\description{
Legend builder for pictograms
}
\keyword{internal}
