% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome.R
\docType{data}
\name{fa5_brand}
\alias{fa5_brand}
\title{Font Awesome 5 Brand}
\format{An object of class \code{character} of length 1.}
\usage{
fa5_brand
}
\description{
`fa5_brand` is shorthand for "`FontAwesome5Brands-Regular`"
}
\keyword{datasets}
