% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome.R
\docType{data}
\name{fa5_fonts}
\alias{fa5_fonts}
\alias{fa5_solid}
\alias{fa5_brand}
\title{Font Awesome 5 easy use font name objets}
\format{An object of class \code{character} of length 1.}
\usage{
fa5_solid

fa5_brand
}
\description{
`fa5_solid` is shorthand for "`FontAwesome5Free-Solid`"

`fa5_brand` is shorthand for "`FontAwesome5Brands-Regular`"
}
\keyword{datasets}
