% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome.R
\docType{data}
\name{fa5_solid}
\alias{fa5_solid}
\title{Font Awesome 5 Solid}
\format{An object of class \code{character} of length 1.}
\usage{
fa5_solid
}
\description{
`fa5_solid` is shorthand for "`FontAwesome5Free-Solid`"
}
\keyword{datasets}
