% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-pictogram.R
\docType{data}
\name{geom_pictogram}
\alias{geom_pictogram}
\alias{GeomPictogram}
\title{Pictogram Geom}
\format{An object of class \code{GeomPictogram} (inherits from \code{GeomText}, \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
geom_pictogram(mapping = NULL, data = NULL, n_rows = 10,
  make_proportional = FALSE, flip = FALSE, ..., na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)

GeomPictogram
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()} or
\code{aes_()}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{ggplot()}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{fortify()} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{n_rows}{how many rows should there be in the waffle chart? default is 10}

\item{make_proportional}{compute proportions from the raw values? (i.e. each
value \code{n} will be replaced with \code{n}/\code{sum(n)}); default is \code{FALSE}.}

\item{flip}{If \code{TRUE}, flip x and y coords. n_rows then becomes n_cols.
Useful to achieve waffle column chart effect. Defaults is \code{FALSE}.}

\item{...}{other arguments passed on to \code{layer()}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{borders()}.}
}
\description{
There are two special/critical \code{aes()} mappings:
\itemize{
\item \code{label} (so the geom knows which column to map the glyphs to)
\item \code{values} (which column you're mapping the filling for the squares with)
}
}
\keyword{datasets}
