% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iron.R
\name{iron}
\alias{iron}
\title{Veritical, left-aligned layout for waffle plots}
\usage{
iron(...)
}
\arguments{
\item{...}{one or more waffle plots}
}
\description{
Uses a technique provided in \url{http://stackoverflow.com/a/13295880/1457051}
to left-align the waffle plots by x-axis. Use the \code{pad} parameter in
\code{waffle} to pad each plot to the max width (num of squares), otherwise
the plots will be scaled.
}

