% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-pictogram.R
\name{scale_label_pictogram}
\alias{scale_label_pictogram}
\title{Used with geom_pictogram() to map Font Awesome fonts to labels}
\usage{
scale_label_pictogram(..., values, aesthetics = "label")
}
\arguments{
\item{...}{dots}

\item{values}{values}

\item{aesthetics}{aesthetics}
}
\description{
Used with geom_pictogram() to map Font Awesome fonts to labels
}
