% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle-enhance.R
\name{theme_enhance_waffle}
\alias{theme_enhance_waffle}
\title{Waffle chart theme cruft remover that can be used with any other theme}
\usage{
theme_enhance_waffle()
}
\description{
Removes:
}
\details{
\itemize{
\item panel grid
\item all axis text
\item all axis ticks
\item all axis titles
}
}
