% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle-package.R
\docType{package}
\name{waffle-package}
\alias{waffle-package}
\title{A package to make waffle charts (square pie charts) in R.}
\description{
For glyphs:\cr
\cr
Font Awesome by Dave Gandy - http://fontawesome.io\cr
License: SIL OFL 1.1\cr
URL: http://scripts.sil.org/OFL
}
