% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontawesome.R
\name{fa_grep}
\alias{fa_grep}
\title{Search Font Awesome glyph names for a pattern}
\usage{
fa_grep(pattern)
}
\arguments{
\item{pattern}{pattern to search for in the names of Font Awesome fonts}
}
\description{
Search Font Awesome glyph names for a pattern
}
