% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle-package.R
\docType{package}
\name{waffle-package}
\alias{waffle-package}
\title{A package to make waffle charts (square pie charts) in R.}
\description{
For glyphs:\cr
\cr
Font Awesome by Dave Gandy - http://fontawesome.io\cr
License: SIL OFL 1.1\cr
URL: http://scripts.sil.org/OFL
}
\author{
\strong{Maintainer}: Bob Rudis \email{bob@rud.is} (\href{https://orcid.org/0000-0001-5670-2640}{ORCID})

Authors:
\itemize{
  \item Dave Gandy (Font Awesome)
}

Other contributors:
\itemize{
  \item Andrew Breza \email{andrew.breza@gmail.com} [contributor]
  \item Matthias Jütte [contributor]
  \item Paul Campbell \email{pacampbell91@gmail.com} (geom_bar flip) [contributor]
}

}
