% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/waffle-package.R
\docType{package}
\name{waffle}
\alias{waffle}
\alias{waffle-package}
\title{A package to make waffle charts (square pie charts)}
\description{
A package to make waffle charts (square pie charts)
}
\author{
Bob Rudis (@hrbrmstr)
}

